use-cases/ai/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Teleport Chat Interface"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "chainlit>=2.9.3",
    "httpx>=0.28.1",
    "PyJWT>=2.9.0",
]
